fix: Fixed the bottom navigation's animation in Nearby tab.
Description:
When switching Nearby tab and its adjacent tab fast, the bottom navigation's y location keeps increasing or decreasing.

Cause:
Given an animation, either the sliding up or down, the animation stats with the bottom navigation's y location of the view currently animated.

Solution:
Set a initial value to the view when an animation starts so that the sliding up animation begins with the view on the bottom and the sliding down animation begins with the view at the y location the bottom navigation fully expanded.