fix: Fixed that child fragments are not visible when the up action is performed
Description:
When a transition from MainFragment to NotificationFragment or SearchingFragment takes place and then the up action is performed, a page fragment in MainFragment (that was visible before the transition) will not be visible until the page is switched.

Cause:
Life-cycles of the fragments in ViewPager is not connected to the activity containing it, namely, the page fragments in ViewPager are contained in MainFragment contained in the MainActivity

Solution:
Replaced the fragment manager passed to HomeFragmentAdapter to a child fragment manager.